restructure fsspec files
