+++ unstructured/ingest/cli/cmds/__init__.py
@@ -6,23 +6,25 @@ import typing as t
 from unstructured.ingest.cli.base.src import BaseSrcCmd
 
 from .airtable import get_base_src_cmd as airtable_base_src_cmd
-from .azure import get_base_dest_cmd as azure_base_dest_cmd
-from .azure import get_base_src_cmd as azure_base_src_cmd
 from .azure_cognitive_search import get_base_dest_cmd as azure_cognitive_search_base_dest_cmd
 from .biomed import get_base_src_cmd as biomed_base_src_cmd
-from .box import get_base_dest_cmd as box_base_dest_cmd
-from .box import get_base_src_cmd as box_base_src_cmd
 from .confluence import get_base_src_cmd as confluence_base_src_cmd
 from .delta_table import get_base_dest_cmd as delta_table_dest_cmd
 from .delta_table import get_base_src_cmd as delta_table_base_src_cmd
 from .discord import get_base_src_cmd as discord_base_src_cmd
-from .dropbox import get_base_dest_cmd as dropbox_base_dest_cmd
-from .dropbox import get_base_src_cmd as dropbox_base_src_cmd
 from .elasticsearch import get_base_src_cmd as elasticsearch_base_src_cmd
-from .fsspec import get_base_dest_cmd as fsspec_base_dest_cmd
-from .fsspec import get_base_src_cmd as fsspec_base_src_cmd
-from .gcs import get_base_dest_cmd as gcs_base_dest_cmd
-from .gcs import get_base_src_cmd as gcs_base_src_cmd
+from .fsspec.azure import get_base_dest_cmd as azure_base_dest_cmd
+from .fsspec.azure import get_base_src_cmd as azure_base_src_cmd
+from .fsspec.box import get_base_dest_cmd as box_base_dest_cmd
+from .fsspec.box import get_base_src_cmd as box_base_src_cmd
+from .fsspec.dropbox import get_base_dest_cmd as dropbox_base_dest_cmd
+from .fsspec.dropbox import get_base_src_cmd as dropbox_base_src_cmd
+from .fsspec.fsspec import get_base_dest_cmd as fsspec_base_dest_cmd
+from .fsspec.fsspec import get_base_src_cmd as fsspec_base_src_cmd
+from .fsspec.gcs import get_base_dest_cmd as gcs_base_dest_cmd
+from .fsspec.gcs import get_base_src_cmd as gcs_base_src_cmd
+from .fsspec.s3 import get_base_dest_cmd as s3_base_dest_cmd
+from .fsspec.s3 import get_base_src_cmd as s3_base_src_cmd
 from .github import get_base_src_cmd as github_base_src_cmd
 from .gitlab import get_base_src_cmd as gitlab_base_src_cmd
 from .google_drive import get_base_src_cmd as google_drive_base_src_cmd
@@ -35,8 +37,6 @@ from .onedrive import get_base_src_cmd as onedrive_base_src_cmd
 from .outlook import get_base_src_cmd as outlook_base_src_cmd
 from .pinecone import get_base_dest_cmd as pinecone_base_dest_cmd
 from .reddit import get_base_src_cmd as reddit_base_src_cmd
-from .s3 import get_base_dest_cmd as s3_base_dest_cmd
-from .s3 import get_base_src_cmd as s3_base_src_cmd
 from .salesforce import get_base_src_cmd as salesforce_base_src_cmd
 from .sharepoint import get_base_src_cmd as sharepoint_base_src_cmd
 from .slack import get_base_src_cmd as slack_base_src_cmd
