+++ unstructured/ingest/connector/fsspec/s3.py
@@ -2,7 +2,7 @@ import typing as t
 from dataclasses import dataclass
 from typing import Type
 
-from unstructured.ingest.connector.fsspec import (
+from unstructured.ingest.connector.fsspec.fsspec import (
     FsspecDestinationConnector,
     FsspecIngestDoc,
     FsspecSourceConnector,
@@ -56,3 +56,7 @@ class S3SourceConnector(FsspecSourceConnector):
 class S3DestinationConnector(FsspecDestinationConnector):
     connector_config: SimpleS3Config
     write_config: S3WriteConfig
+
+    @requires_dependencies(["s3fs", "fsspec"], extras="s3")
+    def initialize(self):
+        super().initialize()
