+++ unstructured/ingest/connector/registry.py
@@ -2,18 +2,19 @@ import json
 from typing import Dict, Type, cast
 
 from unstructured.ingest.connector.airtable import AirtableIngestDoc
-from unstructured.ingest.connector.azure import AzureBlobStorageIngestDoc
 from unstructured.ingest.connector.biomed import BiomedIngestDoc
-from unstructured.ingest.connector.box import BoxIngestDoc
 from unstructured.ingest.connector.confluence import ConfluenceIngestDoc
 from unstructured.ingest.connector.delta_table import DeltaTableIngestDoc
 from unstructured.ingest.connector.discord import DiscordIngestDoc
-from unstructured.ingest.connector.dropbox import DropboxIngestDoc
 from unstructured.ingest.connector.elasticsearch import (
     ElasticsearchIngestDoc,
     ElasticsearchIngestDocBatch,
 )
-from unstructured.ingest.connector.gcs import GcsIngestDoc
+from unstructured.ingest.connector.fsspec.azure import AzureBlobStorageIngestDoc
+from unstructured.ingest.connector.fsspec.box import BoxIngestDoc
+from unstructured.ingest.connector.fsspec.dropbox import DropboxIngestDoc
+from unstructured.ingest.connector.fsspec.gcs import GcsIngestDoc
+from unstructured.ingest.connector.fsspec.s3 import S3IngestDoc
 from unstructured.ingest.connector.github import GitHubIngestDoc
 from unstructured.ingest.connector.gitlab import GitLabIngestDoc
 from unstructured.ingest.connector.google_drive import GoogleDriveIngestDoc
@@ -27,7 +28,6 @@ from unstructured.ingest.connector.notion.connector import (
 from unstructured.ingest.connector.onedrive import OneDriveIngestDoc
 from unstructured.ingest.connector.outlook import OutlookIngestDoc
 from unstructured.ingest.connector.reddit import RedditIngestDoc
-from unstructured.ingest.connector.s3 import S3IngestDoc
 from unstructured.ingest.connector.salesforce import SalesforceIngestDoc
 from unstructured.ingest.connector.sharepoint import SharepointIngestDoc
 from unstructured.ingest.connector.slack import SlackIngestDoc
