+++ unstructured/ingest/runner/fsspec/box.py
@@ -6,7 +6,7 @@ from unstructured.ingest.runner.base_runner import Runner
 from unstructured.ingest.runner.utils import update_download_dir_remote_url
 
 if t.TYPE_CHECKING:
-    from unstructured.ingest.connector.box import SimpleBoxConfig
+    from unstructured.ingest.connector.fsspec.box import SimpleBoxConfig
 
 
 class BoxRunner(Runner):
@@ -25,7 +25,7 @@ class BoxRunner(Runner):
             logger=logger,
         )
 
-        from unstructured.ingest.connector.box import BoxSourceConnector
+        from unstructured.ingest.connector.fsspec.box import BoxSourceConnector
 
         source_doc_connector = BoxSourceConnector(  # type: ignore
             read_config=self.read_config,
