+++ unstructured/ingest/runner/fsspec/gcs.py
@@ -7,7 +7,7 @@ from unstructured.ingest.runner.base_runner import Runner
 from unstructured.ingest.runner.utils import update_download_dir_remote_url
 
 if t.TYPE_CHECKING:
-    from unstructured.ingest.connector.gcs import SimpleGcsConfig
+    from unstructured.ingest.connector.fsspec.gcs import SimpleGcsConfig
 
 
 @dataclass
@@ -23,6 +23,6 @@ class GCSRunner(Runner):
         )
 
     def get_source_connector_cls(self) -> t.Type[BaseSourceConnector]:
-        from unstructured.ingest.connector.gcs import GcsSourceConnector
+        from unstructured.ingest.connector.fsspec.gcs import GcsSourceConnector
 
         return GcsSourceConnector
